requirements.txt
black==24.4.2
flake8==7.0.0
fsspec==2025.3.2
huggingface-hub==0.31.1
isort==5.13.2
kaggle==1.5.16
matplotlib>=3.4.0
numpy==2.2.5
opencv-python==4.9.0.80
packaging==25.0
pandas>=1.3.0
pillow==11.2.1
psutil==7.0.0
python-json-logger==3.3.0
pyyaml==6.0.2
requests==2.32.3
safetensors==0.5.3
scipy==1.15.3
seaborn>=0.11.0
setuptools==80.3.1
tensorboard==2.19.0
timm==1.0.15
torch==2.7.0+cu126
torchvision==0.22.0
tqdm==4.67.1
torch-pruning==1.5.2
torch-summary==1.4.5
torchinfo==1.8.0
wandb==0.17.0